[top_level,dpi] Remove implementation details from header files
