+++ hw/dv/dpi/uartdpi/uartdpi.h
@@ -9,21 +9,12 @@
 extern "C" {
 #endif
 
-#include <stdio.h>
-
-struct uartdpi_ctx {
-  char ptyname[64];
-  int host;
-  int device;
-  char tmp_read;
-  FILE *log_file;
-};
-
 void *uartdpi_create(const char *name, const char *log_file_path);
 void uartdpi_close(void *ctx_void);
 int uartdpi_can_read(void *ctx_void);
 char uartdpi_read(void *ctx_void);
 void uartdpi_write(void *ctx_void, char c);
+
 #ifdef __cplusplus
 }  // extern "C"
 #endif
